+++ CHANGELOG.md
@@ -66,6 +66,7 @@ GRDB adheres to [Semantic Versioning](https://semver.org/), with one exception: 
 ### New
 
 - [#706](https://github.com/groue/GRDB.swift/pull/706): Enhance SQLLiteral and SQL interpolation again
+- [#710](https://github.com/groue/GRDB.swift/pull/710): Check if all migrations have been applied
 - [#712](https://github.com/groue/GRDB.swift/pull/712) by [@pakko972](https://github.com/pakko972): Automatic iOS memory management
 - [#713](https://github.com/groue/GRDB.swift/pull/713): Enhance DatabaseMigrator isolation